refactor: SortDirection 위치 변경에 따른 변화 커밋
