+++ src/main/java/com/codeit/findex/indexinfo/controller/IndexInfoController.java
@@ -5,7 +5,7 @@ import com.codeit.findex.indexinfo.dto.IndexInfoCreateRequest;
 import com.codeit.findex.indexinfo.dto.IndexInfoDto;
 import com.codeit.findex.indexinfo.dto.IndexInfoSummaryDto;
 import com.codeit.findex.indexinfo.dto.IndexInfoUpdateRequest;
-import com.codeit.findex.indexinfo.enums.SortDirection;
+import com.codeit.findex.common.enums.SortDirection;
 import com.codeit.findex.indexinfo.service.IndexInfoService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
