+++ src/main/java/com/codeit/findex/indexinfo/repository/IndexInfoQueryRepository.java
@@ -1,7 +1,7 @@
 package com.codeit.findex.indexinfo.repository;
 
 import com.codeit.findex.indexinfo.entity.IndexInfo;
-import com.codeit.findex.indexinfo.enums.SortDirection;
+import com.codeit.findex.common.enums.SortDirection;
 
 import java.util.List;
